+++ src/notification/NotificationManager.h
@@ -218,6 +218,9 @@ private:
     // Called whenever the position provider finds that pressure altitude and
     // geometric altitude differ by implausible amounts, or clears that status.
     void onPressureAltitudeImplausible();
+    
+    // Called whenever the list of oversized aviation maps changes
+    void onOversizedMapsChanged();
 
     // Called whenever the traffic receiver reports a runtime error, or clears
     // the error status.
@@ -230,6 +233,10 @@ private:
     // When notifications for maps and data are temporarily not possible, then
     // use this timer to notify again.
     QTimer mapsAndDataNotificationTimer;
+
+    // Notification for oversized aviation maps; this pointer guards against
+    // showing the notification more than once.
+    QPointer<Notifications::Notification> m_oversizedMapNotification;
 };